Replace brunch build with npm deploy script
Phoenix ships a `package.json` that has a script called `deploy` which
can be used as a generic way to compile your assets during a deploy
instead of using brunch directly.

This change allows an application to change the `deploy` npm script
instead of having to provide a custom compile script.

Here's the template `package.json` Phoenix ships with.
https://github.com/phoenixframework/phoenix/blob/master/installer/templates/static/brunch/package.json