Update default Node from 5 to 6 LTS
One issue with the current configuration is Node 5 was never an LTS release, so moving this up to 6.9.2 brings it to LTS and the most applicable support for Phoenix applications.

The other major issue is that this release relies on NPM 2, which has a tiered module management system, causing some unwieldy and memory-intensive node applications. NPM 3 has a flat listing of node modules, which significantly reduces the size of some applications.

I discovered this change after trying to use the latest Babel, and running into issues deploying my application to Heroku. Once I split off of this config and opted for LTS, my application was able to deploy and my application size was down quite a bit.